Alkousa/requirements.txt
pygcn
python-telegram-bot[socks]
urllib3[socks]
